Fixing Project routes
creation routes for project fixed
