+++ server.js
@@ -44,7 +44,6 @@ const gformRouter = require('./routes/gform')
 const eventRouter = require('./routes/events')
 const registerRouter = require('./routes/register')
 const adminRouter = require('./routes/admin')
-const projectRouter = require('./routes/project')
 // default login page can pass alerts to it on load
 // hence can notify user easily of failed login routes
 app.get('/', (req, res) => {
@@ -80,7 +79,6 @@ app.use('/gform', gformRouter)
 app.use('/events', eventRouter)
 app.use('/register', registerRouter)
 app.use('/admin', adminRouter)
-app.use('/project', projectRouter)
 app.listen(port, () => {
   console.log(`listening on port : ${port}`)
 })
