+++ views/admin_landing.ejs
@@ -67,6 +67,7 @@
                         <a href="/admin/profile/update/<%=id%>" class="btn btn-outline-primary">Update Profile</a>
                         <a href="/admin/clubs/retrieve" class="btn btn-outline-danger">View clubs</a>
                         <a href="/admin/create_club/<%=id%>" class="btn btn-outline-warning">Create club</a>
+                        <a href="/admin/create_project/<%=id%>" class="btn btn-outline-warning">Create Project</a>
                         <a href="/admin/achievement/create/" class="btn btn-outline-success">Achievement section</a>
                     </form>
