+++ views/create_project.ejs
@@ -1,0 +1,144 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css" integrity="sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk" crossorigin="anonymous">
+    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
+    <title>Add Project</title>
+  </head>
+  <body>
+     <div>
+        <form method="POST" style="position: fixed; left: 0; top:0; padding: 10px;" action="/admin/" >
+                 <button type="submit" class='btn btn-outline-info btn-sm'>HOME</button> 
+            </form>
+        </div>
+ 
+    <h3 style="text-align: center; text-decoration: underline; padding-top: 30px;">Add New Project</h3>
+
+    <div class="container" style="padding-top:30px;">
+        <div class="col-md-5 m-auto">
+
+    <form enctype='text/json' method="POST" action="/admin/project/create">
+        <div class="form-group">
+            <label>Title</label>
+            <input class="form-control" name="title" placeholder="Enter Title of Project" required/>
+        </div>
+            <div class="container">
+                <div class="row clearfix">
+                    <div class="col-md-12 column">
+                        <table class="table table-bordered table-hover" id="tab_logic">
+                            <thead>
+                                <tr>
+                                    <th class="text-center">
+                                        #
+                                    </th>
+                                    <th class="text-center">
+                                        Name
+                                    </th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                <tr id='addr0'>
+                                    <td>1</td>
+                                    <td>
+                                    <input type="text" name='team_member[]'  placeholder='Enter Team Members Name' class="form-control"/>
+                                    </td>
+                                </tr>
+                                <tr id='addr1'></tr>
+                            </tbody>
+                        </table>
+                    </div>
+                </div>
+                <button type="button" id="add_row" class="btn btn-default pull-left">Add Row</button><button id='delete_row' class="pull-right btn btn-default">Delete Row</button>
+            </div>
+            <div class="form-group">
+                <label>Description</label>
+                <textarea class="form-control" rows="3"  name="description" placeholder="Describe the project"></textarea>
+            </div>
+
+            <div class="form-group">
+            <label>Branch</label>
+            <input class="form-control" name="branch" placeholder="Enter Branch"/>
+            </div>
+            <div class="form-group">
+                <label>Club</label>
+                <input class="form-control" name="club" placeholder="Enter Club"/>
+            </div>
+            <div class="form-group">
+                <label>Degree</label>
+                <input class="form-control" name="degree" placeholder="Enter Degree"/>
+            </div>
+            <input style="display: none;" name="id" value="<%=id%>"/>
+            <div class="container">
+                <div class="row clearfix">
+                    <div class="col-md-12 column">
+                        <table class="table table-bordered table-hover" id="tab_logic2">
+                            <thead>
+                                <tr>
+                                    <th class="text-center">
+                                        #
+                                    </th>
+                                    <th class="text-center">
+                                        snapshot_url
+                                    </th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                <tr id='addr20'>
+                                    <td>1</td>
+                                    <td>
+                                    <input type="text" name='snapshot_url[]'  placeholder='Input snapshot url' class="form-control"/>
+                                    </td>
+                                </tr>
+                                <tr id='addr21'></tr>
+                            </tbody>
+                        </table>
+                    </div>
+                </div>
+                <button type="button" id="add_row2" class="btn btn-default pull-left">Add Row</button><button id='delete_row2' class="pull-right btn btn-default">Delete Row</button>
+            </div>
+        <button type="submit" class="btn btn-primary">Submit</button>
+
+    </form>
+        </div>
+    </div>
+    <script>		
+        $(document).ready(function(){
+              var j=1;
+             $("#add_row2").click(function(){b=j-1;
+              $('#addr2'+j).html($('#addr2'+b).html()).find('td:first-child').html(j+1);
+              $('#tab_logic2').append('<tr id="addr2'+(j+1)+'"></tr>');
+              j++; 
+          });
+             $("#delete_row2").click(function(){
+                 if(j>1){
+                 $("#addr2"+(j-1)).html('');
+                 j--;
+                 }
+             });
+        
+        });
+
+        $(document).ready(function(){
+      var i=1;
+     $("#add_row").click(function(){b=i-1;
+      $('#addr'+i).html($('#addr'+b).html()).find('td:first-child').html(i+1);
+      $('#tab_logic').append('<tr id="addr'+(i+1)+'"></tr>');
+      i++; 
+  });
+     $("#delete_row").click(function(){
+    	 if(i>1){
+		 $("#addr"+(i-1)).html('');
+		 i--;
+		 }
+	 });
+
+});
+
+    </script>
+    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
+    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js" integrity="sha384-OgVRvuATP1z7JjHLkuOU7Xw704+h835Lr+6QL9UvYjZE3Ipu6Tp75j7Bh/kR0JKI" crossorigin="anonymous"></script>
+  </body>
+</html>
